move PQLayout to PQVector
